fix linux build errors

+++ Server/Components/Pawn/Script/Script.cpp
@@ -474,10 +474,10 @@ int AMXAPI amx_Allot(AMX* amx, int cells, cell* amx_addr, cell** phys_addr)
 	if (amx->stk < amx->hea + cells * sizeof(cell))
 	{
 		PawnManager::Get()->core->logLn(LogLevel::Error, "Unable to find enough memory for your data.");
-		PawnManager::Get()->core->logLn(LogLevel::Error, "Size: %d bytes, Available space: %d bytes, Need extra size: %d bytes",
-			amx->hea + cells * sizeof(cell), amx->stk, amx->hea + cells * sizeof(cell) - amx->stk);
-		PawnManager::Get()->core->logLn(LogLevel::Error, "You can increase your available memory size by using `#pragma dynamic %d`.",
-			amx->hea / sizeof(cell) + cells);
+		PawnManager::Get()->core->logLn(LogLevel::Error, "Size: %i bytes, Available space: %i bytes, Need extra size: %i bytes",
+			int(amx->hea + cells * sizeof(cell)), amx->stk, int(amx->hea + cells * sizeof(cell) - amx->stk));
+		PawnManager::Get()->core->logLn(LogLevel::Error, "You can increase your available memory size by using `#pragma dynamic %i`.",
+			int(amx->hea / sizeof(cell) + cells));
 		return AMX_ERR_MEMORY;
 	}
